\name{demoCodeWorkSpace}
\alias{demoCodeWorkSpace}
\alias{demoMemUsage2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Showing off some functionality
}
\description{
To be ignored. Mainly demonstrate functionality.
}
\usage{
demoCodeWorkSpace()
}
%- maybe also 'usage' for other objects documented here.
\author{
Nick Sabbe
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~ignore }
\keyword{ ~demo }
