\name{dfr2mat}
\alias{dfr2mat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Convert data.frame to numerical matrix
}
\description{
As the title says
}
\usage{
dfr2mat(dfr)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dfr}{
data.frame to be converted.
}
}
\value{
matrix with the same dimensions as the data.frame
}
\author{
Nick Sabbe
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
dfr2mat(iris)
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
}
\keyword{ ~data.frame }
\keyword{ ~matrix }
