\name{newLine}
\alias{newLine}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Go to new line in output
}
\description{
Go to new line in output - defined more for conformity with other languages.
}
\usage{
newLine()
}
%- maybe also 'usage' for other objects documented here.
\value{
\code{invisible("\n")}
}
\author{
Nick Sabbe
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
cat("test")
newLine()
}

